# cmd/app/main.go
package main

import (
	"database/sql"
	"encoding/json"
	"log"
	"net/http"
	"os"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"
	"github.com/joho/godotenv"
	_ "github.com/lib/pq"
	httpSwagger "github.com/swaggo/http-swagger"
	"golang.org/x/crypto/bcrypt"

	authHttp "lms_backend/internal/auth/delivery/http"
	authMiddleware "lms_backend/internal/auth/delivery/middleware"
	"lms_backend/internal/auth/repository"
	authUseCase "lms_backend/internal/auth/usecase"

	dashboardHttp "lms_backend/internal/dashboard/delivery/http"
	dashboardRepo "lms_backend/internal/dashboard/repository"
	dashboardUseCase "lms_backend/internal/dashboard/usecase"

	contentAdminHttp "lms_backend/internal/content_admin/delivery/http"
	contentAdminRepo "lms_backend/internal/content_admin/repository"
	contentAdminUseCase "lms_backend/internal/content_admin/usecase"

	learningHttp "lms_backend/internal/learning/delivery/http"
	learningRepo "lms_backend/internal/learning/repository"
	learningUseCase "lms_backend/internal/learning/usecase"

	reviewHttp "lms_backend/internal/review/delivery/http"
	reviewRepo "lms_backend/internal/review/repository"
	reviewUseCase "lms_backend/internal/review/usecase"

	profileHttp "lms_backend/internal/profile/delivery/http"
	profileRepo "lms_backend/internal/profile/repository"
	profileUseCase "lms_backend/internal/profile/usecase"

	scheduleHttp "lms_backend/internal/schedule/delivery/http"
	scheduleRepo "lms_backend/internal/schedule/repository"
	scheduleUseCase "lms_backend/internal/schedule/usecase"

	chatHttp "lms_backend/internal/chat/delivery/http"
	chatRepo "lms_backend/internal/chat/repository"
	chatUseCase "lms_backend/internal/chat/usecase"

	"lms_backend/internal/domain"
	dbPkg "lms_backend/pkg/database"
	storageService "lms_backend/pkg/storage"
)

// @title Cap Education LMS - API
// @version 1.0
// @description API для LMS платформы Cap Education.
// @host capapi.grtsq.ru
// @BasePath /
func main() {
	if err := godotenv.Load(); err != nil {
		log.Println("No .env file found (using system envs)")
	}

	apiPort := os.Getenv("API_PORT")
	if apiPort == "" {
		apiPort = "8080"
	}

	host := os.Getenv("DB_HOST")
	port := os.Getenv("DB_PORT")
	user := os.Getenv("DB_USER")
	password := os.Getenv("DB_PASSWORD")
	dbname := os.Getenv("DB_NAME")

	connStr := "host=" + host + " port=" + port + " user=" + user +
		" password=" + password + " dbname=" + dbname + " sslmode=disable"

	db, err := sql.Open("postgres", connStr)
	if err != nil {
		log.Fatal(err)
	}
	defer db.Close()
	log.Println("Успешное подключение к PostgreSQL")

	if err := dbPkg.RunMigrations(db); err != nil {
		log.Fatalf("Critical Error running migrations: %v", err)
	}

	s3Client, err := storageService.NewS3Client(
		os.Getenv("S3_ENDPOINT_URL"),
		os.Getenv("S3_REGION"),
		os.Getenv("S3_BUCKET_NAME"),
		os.Getenv("S3_ACCESS_KEY_ID"),
		os.Getenv("S3_SECRET_ACCESS_KEY"),
	)
	if err != nil {
		log.Fatalf("Failed to initialize S3 Storage: %v", err)
	}

	authRepoImpl := repository.NewAuthRepository(db)
	authUsecase := authUseCase.NewAuthUsecase(authRepoImpl)
	authHandler := authHttp.NewAuthHandler(authUsecase)

	dashboardRepository := dashboardRepo.NewDashboardRepository(db)
	dashboardUsecase := dashboardUseCase.NewDashboardUseCase(dashboardRepository)
	dashboardHandler := dashboardHttp.NewDashboardHandler(dashboardUsecase)

	adminRepo := contentAdminRepo.NewContentAdminRepository(db)
	adminUsecase := contentAdminUseCase.NewContentAdminUseCase(adminRepo, s3Client)
	adminHandler := contentAdminHttp.NewContentAdminHandler(adminUsecase)

	learningRepoImpl := learningRepo.NewLearningRepository(db)
	learningUC := learningUseCase.NewLearningUseCase(learningRepoImpl, s3Client)
	learningHandler := learningHttp.NewLearningHandler(learningUC)

	reviewRepoImpl := reviewRepo.NewReviewRepository(db)
	reviewUC := reviewUseCase.NewReviewUseCase(reviewRepoImpl)
	reviewHandler := reviewHttp.NewReviewHandler(reviewUC)

	profileRepoImpl := profileRepo.NewProfileRepository(db)
	profileUC := profileUseCase.NewProfileUseCase(profileRepoImpl, s3Client)
	profileHandler := profileHttp.NewProfileHandler(profileUC)

	scheduleRepoImpl := scheduleRepo.NewScheduleRepository(db)
	scheduleUC := scheduleUseCase.NewScheduleUseCase(scheduleRepoImpl)
	scheduleHandler := scheduleHttp.NewScheduleHandler(scheduleUC)

	chatRepoImpl := chatRepo.NewChatRepository(db)
	chatUC := chatUseCase.NewChatUseCase(chatRepoImpl)
	chatHandler := chatHttp.NewChatHandler(chatUC)

	r := chi.NewRouter()

	r.Use(cors.Handler(cors.Options{
		AllowedOrigins: []string{
			"http://localhost:3000",
			"https://localhost:3000",
			"http://test.xcx.cx",
			"https://test.xcx.cx",
			"http://capapi.grtsq.ru",
			"https://capapi.grtsq.ru",
			"https://cap.grtsq.ru",
			"https://cap-education.vercel.app",
		},
		AllowedMethods:   []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
		AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type", "X-CSRF-Token", "Cookie", "X-System-Secret"},
		ExposedHeaders:   []string{"Link"},
		AllowCredentials: true,
		MaxAge:           300,
	}))

	r.Use(middleware.Logger)
	r.Use(middleware.Recoverer)

	r.Post("/auth/register", authHandler.Register)
	r.Post("/auth/login", authHandler.Login)
	r.Post("/auth/logout", authHandler.Logout)

	r.Post("/system/reset-password", func(w http.ResponseWriter, r *http.Request) {
		secret := os.Getenv("SYSTEM_SECRET")
		if secret == "" || r.Header.Get("X-System-Secret") != secret {
			http.Error(w, "Forbidden", http.StatusForbidden)
			return
		}

		var body struct {
			Email    string `json:"email"`
			Password string `json:"password"`
		}
		if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
			http.Error(w, "Bad Request", http.StatusBadRequest)
			return
		}

		hash, err := bcrypt.GenerateFromPassword([]byte(body.Password), 12)
		if err != nil {
			http.Error(w, "Hash Error", http.StatusInternalServerError)
			return
		}

		res, err := db.Exec("UPDATE users SET password_hash=$1 WHERE email=$2", string(hash), body.Email)
		if err != nil {
			http.Error(w, "DB Error: "+err.Error(), http.StatusInternalServerError)
			return
		}

		affected, _ := res.RowsAffected()
		if affected == 0 {
			_, err = db.Exec(`INSERT INTO users (id, first_name, last_name, email, password_hash, role)
				VALUES (gen_random_uuid(), 'Root', 'Admin', $1, $2, 'admin')`, body.Email, string(hash))
			if err != nil {
				http.Error(w, "Create Error: "+err.Error(), http.StatusInternalServerError)
				return
			}
			w.WriteHeader(http.StatusCreated)
			w.Write([]byte("User Created and Password Set"))
			return
		}

		w.WriteHeader(http.StatusOK)
		w.Write([]byte("Password Updated Successfully"))
	})

	r.Group(func(r chi.Router) {
		r.Use(authMiddleware.AuthMiddleware, authMiddleware.RoleRequiredMiddleware(domain.RoleAdmin, domain.RoleTeacher, domain.RoleModerator, domain.RoleCurator))

		r.Get("/admin/dashboard/stats", dashboardHandler.GetAdminDashboard)
		r.Get("/admin/courses", adminHandler.GetAllCourses)
		r.Post("/admin/courses", adminHandler.CreateCourse)
		r.Put("/admin/courses/{id}/settings", adminHandler.UpdateCourseSettings)
		r.Get("/admin/courses/{id}/structure", adminHandler.GetCourseStructure)
		r.Get("/admin/courses/{id}/students", adminHandler.GetCourseStudents)
		r.Get("/admin/courses/{id}/stats", adminHandler.GetCourseStats)

		r.Post("/admin/modules", adminHandler.CreateModule)
		r.Delete("/admin/modules/{id}", adminHandler.DeleteModule)
		r.Get("/admin/tests/{id}", adminHandler.GetTest)
        r.Get("/admin/projects/{id}", adminHandler.GetProject)

		r.Post("/admin/lessons", adminHandler.CreateLesson)
		r.Get("/admin/lessons/{id}", adminHandler.GetLesson)    
		r.Put("/admin/lessons/{id}", adminHandler.UpdateLesson)
		r.Delete("/admin/lessons/{id}", adminHandler.DeleteLesson)
		r.Post("/admin/modules/bulk", adminHandler.CreateModulesBulk)
		r.Post("/admin/lessons/bulk", adminHandler.CreateLessonsBulk)
		r.Post("/admin/tests", adminHandler.CreateTest)
		r.Delete("/admin/tests/{id}", adminHandler.DeleteTest)

		r.Post("/admin/projects", adminHandler.CreateProject)
		r.Delete("/admin/projects/{id}", adminHandler.DeleteProject)

		r.Post("/admin/media/upload", adminHandler.UploadMedia)

		r.Get("/admin/users", adminHandler.GetUsersList)
		r.Get("/admin/users/{id}", adminHandler.GetUserInfo)
		r.Post("/admin/users", adminHandler.CreateUser)
		r.Put("/admin/users/{id}", adminHandler.UpdateUser)
		r.Delete("/admin/users/{id}", adminHandler.DeleteUser)
		r.Post("/admin/enroll", adminHandler.EnrollUser)
		r.Delete("/admin/courses/{id}/enroll/{user_id}", adminHandler.UnenrollStudent)

		r.Get("/admin/users/all", adminHandler.GetAllUsersTable)
		r.Get("/admin/students/detailed", adminHandler.GetDetailedStudents)
		r.Get("/admin/teachers/detailed", adminHandler.GetDetailedTeachers)
		r.Get("/admin/curators/detailed", adminHandler.GetDetailedCurators)
		r.Get("/admin/moderators/detailed", adminHandler.GetDetailedModerators)

		r.Post("/admin/streams", adminHandler.CreateStream)
		r.Get("/admin/streams", adminHandler.GetStreams)
		r.Post("/admin/groups", adminHandler.CreateGroup)
		r.Get("/admin/groups", adminHandler.GetGroups)



		r.Get("/staff/submissions", reviewHandler.GetPendingSubmissions)
		r.Post("/staff/submissions/{id}/evaluate", reviewHandler.EvaluateSubmission)
	})

	r.Group(func(r chi.Router) {
		r.Use(authMiddleware.AuthMiddleware)
		r.Get("/teachers", learningHandler.GetTeachers)
		r.Get("/teachers/{id}", learningHandler.GetTeacherDetails)
		r.Post("/teachers/{id}/reviews", learningHandler.AddReview)
		r.Get("/teacher/profile", learningHandler.GetTeacherDashboard)
		r.Get("/dashboard/home", dashboardHandler.GetUserHome)
		r.Get("/my-courses", learningHandler.GetMyCourses)
		r.Get("/courses/{id}", learningHandler.GetCourseContent)
		r.Get("/lessons/{id}", learningHandler.GetLessonDetail)
		r.Post("/lessons/{id}/assignment", learningHandler.SubmitAssignment)
		r.Post("/lessons/{id}/complete", learningHandler.CompleteLesson)
		r.Post("/admin/courses/bulk", adminHandler.CreateFullCourse)
		r.Get("/tests/{id}", learningHandler.GetTest)
        r.Get("/projects/{id}", learningHandler.GetProject)

		r.Get("/profile", profileHandler.GetProfile)
		r.Put("/profile", profileHandler.UpdateProfile)
		r.Put("/profile/teacher/schedule", profileHandler.UpdateTeacherSchedule)

		r.Get("/schedule/weekly", scheduleHandler.GetWeeklySchedule)
		r.Get("/schedule/monthly", scheduleHandler.GetMonthlySchedule)

		r.Get("/chat/ws", chatHandler.ConnectToChat)
		r.Get("/chat/history", chatHandler.GetChatHistory)
	})

	r.Get("/swagger/*", httpSwagger.Handler(
		httpSwagger.URL("/docs/swagger.json"), 
	))
	r.Get("/docs/*", func(w http.ResponseWriter, r *http.Request) {
		http.StripPrefix("/docs/", http.FileServer(http.Dir("./docs"))).ServeHTTP(w, r)
	})

	listenAddr := ":" + apiPort
	log.Printf("Server started on port: %s", apiPort)

	if err := http.ListenAndServe(listenAddr, r); err != nil {
		log.Fatalf("Server failed to start: %v", err)
	}
}