# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для LMS платформы Cap Education.",
        "title": "Cap Education LMS - API",
        "contact": {},
        "version": "1.0"
    },
    "host": "capapi.grtsq.ru",
    "basePath": "/",
    "paths": {
        "/admin/courses": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Список всех курсов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Course"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Создание нового курса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название курса",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Описание курса",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Флаг основного курса (true/false)",
                        "name": "is_main",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Изображение обложки курса",
                        "name": "image_file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/courses/bulk": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Создание курса со всей структурой (Модули + Уроки)",
                "parameters": [
                    {
                        "description": "Структура курса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateBulkCourseInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/courses/{id}/settings": {
            "put": {
                "description": "Редактирование параметров курса (ДЗ, Тесты, Discord) и загрузка новой обложки.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Обновление настроек курса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID курса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название курса",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Описание курса",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Основной курс",
                        "name": "is_main",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Статус (draft, active, archived)",
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Включить домашние задания",
                        "name": "has_homework",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "ДЗ обязательно для прохода дальше",
                        "name": "is_homework_mandatory",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Тесты обязательны",
                        "name": "is_test_mandatory",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Проекты обязательны",
                        "name": "is_project_mandatory",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Discord обязателен",
                        "name": "is_discord_mandatory",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Защита от копирования",
                        "name": "is_anti_copy_enabled",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Новая обложка курса",
                        "name": "cover_image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/courses/{id}/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Stats"
                ],
                "summary": "ADMIN: Статистика курса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID курса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.AdminCourseStats"
                        }
                    }
                }
            }
        },
        "/admin/courses/{id}/structure": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Содержание курса (Дерево)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID курса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CourseStructure"
                        }
                    }
                }
            }
        },
        "/admin/courses/{id}/students": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Users"
                ],
                "summary": "ADMIN: Ученики конкретного курса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID курса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.AdminStudentProgress"
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/enroll": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Users"
                ],
                "summary": "ADMIN: Запись на курс",
                "parameters": [
                    {
                        "description": "UserID и CourseID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.EnrollRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/groups": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Staff"
                ],
                "summary": "ADMIN: Список групп потока",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID потока (опционально)",
                        "name": "stream_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Group"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Staff"
                ],
                "summary": "ADMIN: Создать группу",
                "parameters": [
                    {
                        "description": "Данные группы",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/lessons": {
            "post": {
                "description": "Создание урока. Можно слать JSON (для блочного контента) или multipart/form-data (для старой загрузки файлов).",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Добавление урока",
                "parameters": [
                    {
                        "description": "Данные урока (JSON)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/http.CreateLessonRequest"
                        }
                    },
                    {
                        "type": "file",
                        "description": "Видео файл (только для multipart)",
                        "name": "video_file",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Презентация (только для multipart)",
                        "name": "presentation_file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/lessons/bulk": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Массовое создание уроков",
                "parameters": [
                    {
                        "description": "Массив уроков",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/usecase.CreateLessonInput"
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/admin/lessons/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Получить данные урока для редактора",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Lesson"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Редактировать урок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные урока",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateLessonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Удаление урока",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/media/upload": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Загрузка медиа-файла",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Файл для загрузки",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "url",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/modules": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Создание модуля",
                "parameters": [
                    {
                        "description": "Данные модуля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateModuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/modules/bulk": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-Content"
                ],
                "summary": "ADMIN: Массовое создание модулей",
                "parameters": [
                    {
                        "description": "Массив модулей",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/usecase.CreateModuleInput"
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "domain.Course": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_homework": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_anti_copy_enabled": {
                    "type": "boolean"
                },
                "is_discord_mandatory": {
                    "type": "boolean"
                },
                "is_homework_mandatory": {
                    "type": "boolean"
                },
                "is_main": {
                    "type": "boolean"
                },
                "is_project_mandatory": {
                    "type": "boolean"
                },
                "is_test_mandatory": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/domain.CourseStatus"
                },
                "teacher_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "http.EnrollRequest": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "string",
                    "example": "c1111111-1111-1111-1111-111111111111"
                },
                "user_id": {
                    "type": "string",
                    "example": "a0000000-0000-0000-0000-000000000001"
                }
            }
        },
...
        "usecase.CreateLessonInput": {
            "type": "object"
        },
...
    }
}